+++ src/Resources/ResourceManager/SdkClient/ResourceManagerSdkClient.cs
@@ -20,6 +20,7 @@ using System.IO;
 using System.Linq;
 using System.Management.Automation;
 using System.Net;
+using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
@@ -261,14 +262,15 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             var deploymentExtended =  this.WaitDeploymentStatus(
                 getDeploymentFunc,
                 writeProgressAction,
-                deploymentOperationError,
                 ProvisioningState.Canceled,
                 ProvisioningState.Succeeded,
                 ProvisioningState.Failed);
 
             if (deploymentOperationError.ErrorMessages.Count > 0)
             {
-                WriteError(deploymentOperationError.GetErrorMessagesWithOperationId(parameters.DeploymentName));
+                WriteError(GetDeploymentErrorMessagesWithOperationId(deploymentOperationError, 
+                    parameters.DeploymentName, 
+                    deploymentExtended?.Properties?.CorrelationId));
             }
 
             return deploymentExtended;
@@ -317,7 +319,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
         private DeploymentExtended WaitDeploymentStatus(
             Func<Task<AzureOperationResponse<DeploymentExtended>>> getDeployment,
             Action listDeploymentOperations,
-            DeploymentOperationErrorInfo deploymentOperationError,
             params ProvisioningState[] status)
         {
             DeploymentExtended deployment;
@@ -348,10 +349,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
 
                 using (var getResult = getDeploymentTask.ConfigureAwait(false).GetAwaiter().GetResult())
                 {
-                    deploymentOperationError.SetRequestIdFromResponseHeaders(getResult.Response);
-
                     deployment = getResult.Body;
                     var response = getResult.Response;
+
                     if (response != null && response.Headers.RetryAfter != null && response.Headers.RetryAfter.Delta.HasValue)
                     {
                         step = response.Headers.RetryAfter.Delta.Value.Seconds;
@@ -1699,6 +1699,35 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
 
             return null;
+        }
+
+        public string GetDeploymentErrorMessagesWithOperationId(DeploymentOperationErrorInfo errorInfo, string deploymentName = null, string correlationId = null)
+        {
+            if (errorInfo.ErrorMessages.Count == 0)
+                return String.Empty;
+
+            var sb = new StringBuilder();
+
+            int maxErrors = errorInfo.ErrorMessages.Count > DeploymentOperationErrorInfo.MaxErrorsToShow
+               ? DeploymentOperationErrorInfo.MaxErrorsToShow
+               : errorInfo.ErrorMessages.Count;
+
+            // Add outer message showing the total number of errors.
+            sb.AppendFormat(ProjectResources.DeploymentOperationOuterError, deploymentName, maxErrors, errorInfo.ErrorMessages.Count);
+
+            // Add each error message
+            errorInfo.ErrorMessages
+                .Take(maxErrors).ToList()
+                .ForEach(m => sb
+                    .AppendLine()
+                    .AppendFormat(ProjectResources.DeploymentOperationResultError, m
+                            .ToFormattedString())
+                    .AppendLine());
+
+            // Add correlationId
+             sb.AppendLine().AppendFormat(ProjectResources.DeploymentCorrelationId, correlationId);
+
+            return sb.ToString();
         }
     }
 }